Deal with duplicate adopted stylesheets correctly.
This also fixes some other bits, like mAdoptedStyleSheets.ElementAt(0) (which
really should be "the first applicable adopted stylesheet") and so on. Added a
test for that.

The new invariant also implies that a DocumentOrShadowRoot can't appear twice in
StyleSheet::mAdopters.

Also clears adopted stylesheets in Document::ResetStylesheetsToURI, as it's the
right thing to do (if you hit that code path with constructable sheets you'd end
up asserting, as we'd call FillStyleSetDocumentSheets after and find sheets in
there). I don't quite have a test-case for that though.

Differential Revision: https://phabricator.services.mozilla.com/D64151

bugzilla-url: https://bugzilla.mozilla.org/show_bug.cgi?id=1617746
gecko-commit: 6cb30e866b95384c2ebebe10196d05f656716293
gecko-integration-branch: autoland
gecko-reviewers: nordzilla